管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -72,15 +72,16 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void entryNew() throws ServletException, IOException {
+        // 管理者かチェック
+        if (checkAdmin()) {
+            //CSRF対策用トークン
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            // 空の従業員インスタンス
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
-        //CSRF対策用トークン
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        // 空の従業員インスタンス
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
-
-        // 新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
-
+            // 新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -89,39 +90,42 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void create() throws ServletException, IOException {
-        // CSRF対策 tokenのチェック
-        if (checkToken()) {
-            // パラメータの値を元に従業員情報のインスタンスを作成する
-            EmployeeView ev = new EmployeeView(
-                    null,
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+        // 管理者かチェック
+        if (checkAdmin()) {
+            // CSRF対策 tokenのチェック
+            if (checkToken()) {
+                // パラメータの値を元に従業員情報のインスタンスを作成する
+                EmployeeView ev = new EmployeeView(
+                        null,
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            // アプリケーションスコープからpepper文字列を取得
-            String pepper = getContextScope(PropertyConst.PEPPER);
+                // アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            List<String> errors = service.create(ev, pepper);
+                List<String> errors = service.create(ev, pepper);
 
-            if (errors.size() > 0) {
-                // エラーが存在する場合
-                putRequestScope(AttributeConst.TOKEN, getTokenId());
-                putRequestScope(AttributeConst.EMPLOYEE, ev);
-                putRequestScope(AttributeConst.ERR, errors);
+                if (errors.size() > 0) {
+                    // エラーが存在する場合
+                    putRequestScope(AttributeConst.TOKEN, getTokenId());
+                    putRequestScope(AttributeConst.EMPLOYEE, ev);
+                    putRequestScope(AttributeConst.ERR, errors);
 
-                // 新規画面表示
-                forward(ForwardConst.FW_EMP_NEW);
-            } else {
-                // エラーが存在しない場合
-                // セッションに登録完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+                    // 新規画面表示
+                    forward(ForwardConst.FW_EMP_NEW);
+                } else {
+                    // エラーが存在しない場合
+                    // セッションに登録完了のフラッシュメッセージを設定
+                    putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
 
-                // 一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                    // 一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
             }
         }
     }
@@ -132,19 +136,22 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void show() throws ServletException, IOException {
-        // idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        // 管理者かチェック
+        if (checkAdmin()) {
+            // idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            // データが取得出来ない、もしくは論理削除されている場合、エラー画面表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                // データが取得出来ない、もしくは論理削除されている場合、エラー画面表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            // 詳細画面表示
+            forward(ForwardConst.FW_EMP_SHOW);
         }
-
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        // 詳細画面表示
-        forward(ForwardConst.FW_EMP_SHOW);
     }
 
     /**
@@ -153,22 +160,25 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException, IOException {
-        // idを条件に従業員データを取得
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        // 管理者かチェック
+        if (checkAdmin()) {
+            // idを条件に従業員データを取得
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            // データが取得出来なかった、もしくは論理削除済みの場合エラー画面表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                // データが取得出来なかった、もしくは論理削除済みの場合エラー画面表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            // CSRF対策 tokenのチェック
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            // 取得した従業員番号
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            // 変種画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
         }
-
-        // CSRF対策 tokenのチェック
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        // 取得した従業員番号
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        // 変種画面を表示する
-        forward(ForwardConst.FW_EMP_EDIT);
     }
 
     /**
@@ -177,40 +187,43 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void update() throws ServletException, IOException {
-        // CSRF対策 tokenのチェック
-        if (checkToken()) {
-            // パラメータの値を元に従業員情報のインスタンスを作成する
-            EmployeeView ev = new EmployeeView(
-                    toNumber(getRequestParam(AttributeConst.EMP_ID)),
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+        // 管理者かチェック
+        if (checkAdmin()) {
+            // CSRF対策 tokenのチェック
+            if (checkToken()) {
+                // パラメータの値を元に従業員情報のインスタンスを作成する
+                EmployeeView ev = new EmployeeView(
+                        toNumber(getRequestParam(AttributeConst.EMP_ID)),
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            // アプリケーションスコープからpepper文字列を取得
-            String pepper = getContextScope(PropertyConst.PEPPER);
+                // アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            //従業員情報更新
-            List<String> errors = service.update(ev, pepper);
+                //従業員情報更新
+                List<String> errors = service.update(ev, pepper);
 
-            if (errors.size() > 0) {
-                // 更新中にエラーが発生した場合
-                putRequestScope(AttributeConst.TOKEN, getTokenId());
-                putRequestScope(AttributeConst.EMPLOYEE, ev);
-                putRequestScope(AttributeConst.ERR, errors);
+                if (errors.size() > 0) {
+                    // 更新中にエラーが発生した場合
+                    putRequestScope(AttributeConst.TOKEN, getTokenId());
+                    putRequestScope(AttributeConst.EMPLOYEE, ev);
+                    putRequestScope(AttributeConst.ERR, errors);
 
-                // 編集画面を再表示
-                forward(ForwardConst.FW_EMP_EDIT);
-            } else {
-                // エラーでない場合
-                // セッションに更新完了のフラッシュメッセージを設定
-                putRequestScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
+                    // 編集画面を再表示
+                    forward(ForwardConst.FW_EMP_EDIT);
+                } else {
+                    // エラーでない場合
+                    // セッションに更新完了のフラッシュメッセージを設定
+                    putRequestScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
 
-                // 一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                    // 一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
             }
         }
     }
@@ -221,16 +234,40 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void destroy() throws ServletException, IOException {
-        // CSRF対策 tokenのチェック
-        if (checkToken()) {
-            // idを条件に従業員データを論理削除する
-            service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        // 管理者かチェック
+        if (checkAdmin()) {
+            // CSRF対策 tokenのチェック
+            if (checkToken()) {
+                // idを条件に従業員データを論理削除する
+                service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            // セッションに削除完了のフラッシュメッセージを設定
-            putRequestScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+                // セッションに削除完了のフラッシュメッセージを設定
+                putRequestScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
-            // 一覧画面にリダイレクト
-            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                // 一覧画面にリダイレクト
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+            }
+        }
+    }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+        // セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        // 管理者で無ければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
         }
     }
 }
